Delete test file with newline problems.
The file contains the combinations CRLF LFCR CR and LF, and it looks like git automatically changed them to LF in the last dcommit.

See also http://www.dont-panic.cc/capi/2009/02/16/how-to-force-git-to-consider-a-file-as-binary/

The file is going to be added again in a different issue with an additional .gitattributes file to treat it as binary.

R=jkummerow@chromium.org

Review URL: https://codereview.chromium.org/18080013

git-svn-id: https://v8.googlecode.com/svn/branches/bleeding_edge@15377 ce2b1a6d-e550-0410-aec6-3dcde31c8c00