MIPS64: DO_NOT_SUBMIT: Remove a seemingly broken assert when running on os-x.
On Linux x86_64 CpuFeatures::IsSupported(SSE2) is false, and this
ASSERT does not hit, but it is true on mac os-x. This seems wrong, as
x86_64 definitely has SSE2 mode internally, so it should be true
on both linux & mac.

Another odd thing is that the mips32 code does not hit this code
path when running the same test:

TEST: mjsunit/debug-stepout-recursive-function.js
(one of many that fail on mac for this assert)