MIPS64: fr1 mode 5/5: Convert rest of files to use fr1 mode properly.
This is just converting a few mtc1/mfc1 on odd FP regs to use
mthc1/mfhc1 on the single base FP register. All operations on
FP register-pairs are eliminated.

However, this CL does not change the allocator to start using the
new FP registers. That will come in a later change.

This is a hard-conversion to FR1 mode. To later combine this port with
mips32 will require macro-wrapping these to support either mode. At that
time we will also have to support mips32r1 cpus which do not have these
instructions.