MIPS64: Reorder mthc1/mtc1 to prevent FP register trashing.
An mtc1 in FR=1 mode on mips64 hardware will trash the upper 32-bits
(zero-write) of the destination fpr.

So when using both mtc1 and mthc1, the mtc1 must be used before the
mfhc1.

The manual confirms this behavior:  "The low word in GPR rt is
placed into the low word of FPR fs. If FPRs are 64 bits wide,
bits 63..32 of FPR fs become UNPREDICTABLE."

This bug was not seen on simulator. So I have also fixed the sim
to model this behavior, and verfied that this bug now shows on
the sim (prior to this fix).

TEST=mjsunit/smi-ops.js and about 40 others.