MIPS64: Remove flag track-allocation-sites.
Port r18474 (dd59e7a3)

Original commit message:
The flag has been on in the build for ~9 months, and we aren't likely to turn it off. The only customer of the flag is a set of tests that want to verify transitioning behavior in isolation. This CL removes the flag and updates those tests to get what they want without the flag.