MIPS64: Bugfix for 349874: we incorrectly believe we saw a growing store
Port r19691 (ff000d7b)

Original commit message:
When we set an out of bounds array index, the index might be so large that
it causes the array to go to dictionary mode. It's better to avoid
"learning" that this was a growing store in that case.

This fix also partially reverts a fix for bug 347543, as this fix is
comprehensive and satisfies that repro case as well (partial revert of
v19591).