MIPS64: Add SmiLoadUntag(), SmiLoadScale(), SmiScale(), etc.
Intent is to allow Loading, Untagging, and Shifting/Scaling to
be combined. This should allow common source for mips32 & mips64
for these operations.

Though unifying the ports is no longer a short-term goal, these
changes should improve readability, even if just used on mips64.

Note that the optimizations around these are different for mips32
and mips64, and the intent is to minimize total instructions on
both platforms.

Includes one example usage from Dusan's prior CL.