mips tf: workaround to get add/sub w/overflow working.
I had previously created a specialized VisitBinopOverflow()
to handle allocating a scratch register for the overflow-
checking macros, but something failed in reg alloctor, and
needs to be debugged properly.

In the meantime I used this quick hack to use t8 as scratch.

I had also noticed that CodeGenerator::AddNopForSmiCodeInlining()
was getting called, so I temporarily fixed that to work. Since
the comment was that it was not used on ARM, we check and maybe
fix it for mips.