Set the FPU precision control to double-precision.
The default double precision control of FPU is extended double-precision.
While the number definition for JavaScript is double-precision. We use
the FPU control word to set the doulbe precision and replace the original
solution which store the data to memory and load it again.

This patch also fixes the error that Sunspider 1.0.2 can not run with V8 if
sse2 support is disabled.

BUG=
R=weiliang.lin@intel.com

Review URL: https://codereview.chromium.org/700053003

Patch from Chunyang Dai <chunyang.dai@intel.com>.

git-svn-id: https://v8.googlecode.com/svn/branches/bleeding_edge@25125 ce2b1a6d-e550-0410-aec6-3dcde31c8c00